requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gensim>=4.3.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0 